// chrome-extension/src/entities/function/model/function.types.ts
import type { IEntity } from '@/shared/types/entitiy.types.ts'
import type { ICapabilityEntity } from '@/core/types/capability.types.ts'

export type FunctionType = 'button' | 'standalone' | 'dynamic' | 'automation' | 'scheduler' | 'unknown'

export interface IFunctionEntity<TOrigin extends IEntity = IEntity> extends ICapabilityEntity {
    id: string
    type: FunctionType
    displayName: string
    originEntity: TOrigin
    script?: string | null
}

export type FunctionMetadata = {
    type: FunctionType
    icon: string
}